backend/requirements.txt
streamlit==1.32.0
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
googletrans==4.0.0-rc1
pymongo==4.6.3
Flask==3.0.3
pandas==2.2.2
python-dotenv==1.0.1
